requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# D:\Code\Projects\movies-notifications\main.py: 11
PyPDF2 == 1.26.0

# D:\Code\Projects\movies-notifications\main.py: 1
bs4 == 0.0.1

# D:\Code\Projects\movies-notifications\main.py: 9
certifi == 2020.6.20

# D:\Code\Projects\movies-notifications\main.py: 8
pymongo == 4.0

# D:\Code\Projects\movies-notifications\main.py: 4
python-telegram-bot==20.7

# D:\Code\Projects\movies-notifications\main.py: 21
random_user_agent == 1.0.1

# D:\Code\Projects\movies-notifications\main.py: 2
requests == 2.26.0

# D:\Code\Projects\movies-notifications\main.py: 14,15,16
selenium == 4.0.0

# D:\Code\Projects\movies-notifications\main.py: 18
webdriver_manager == 3.5.2

pymongo[srv] == 4.0

python-dotenv == 1.0.0

fastapi == 0.108.0 